Dealt with issues in the construction/coercion behavior for edgelists introduced in 1.17.0 (reported by Ian Fellows).
Fixed problems related to redundant edge removal for edgelists with <=1 rows.

While I was at it, patched up behavior for zero-row matrices (which had not
really been supported before).  These are now assumed to be edgelists if
not dimensionally symmetric and if no other information is provided, and
they will either fail intentionally with an informative error message if
no n attribute is present, or else produce an empty graph.

The problems for the former were related to dimension loss (some drop=FALSE
statements were omitted), and the latter hadn't been supported (so there
were various test conditions that assumed non-zero row matrices in some
cases).