These are selected changes from the commit reproduced below. The selected changes are as follows: 1) ChangeLog update (not actual removal of tests). 2) DESCRIPTION version bumped. 3) CITATION JoSS URL modified (but to a DOI-based one).
Commit 3df881c027d63c53867fe5860ff48c02e969b4f9

This takes the v1.17.1 release branch and deactivates tests for CRAN purposes - we have endless problems with CRAN deciding that the tests take too long (or otherwise being upset with them) and threatening to pull the package.  This is one of those times.  So I have now turned them all off; they can be reactivated for specific use, before pushing to CRAN.  I also had to update a citation, and remove github-specific files that are now angering CRAN. (I can't even.)  But there should be no user-visible changes in this branch, unless you count the citation URL.  Hopefully.